requirements.txt
torch
transformers
pillow
json
